backend-django/ev2project/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
pymysql
dotenv
djangorestframework_simplejwt